Feature: delete location and services validations

See merge request team-10/storoom!23

+++ src/main/java/ru/itmo/highload/storoom/services/LocationService.java
@@ -4,13 +4,16 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import ru.itmo.highload.storoom.consts.UnitStatus;
 import ru.itmo.highload.storoom.exceptions.ResourceAlreadyExistsException;
 import ru.itmo.highload.storoom.exceptions.ResourceNotFoundException;
 import ru.itmo.highload.storoom.models.DTOs;
 import ru.itmo.highload.storoom.models.LocationEntity;
 import ru.itmo.highload.storoom.models.OwnerEntity;
+import ru.itmo.highload.storoom.models.UnitEntity;
 import ru.itmo.highload.storoom.repositories.LocationRepo;
 import ru.itmo.highload.storoom.repositories.OwnerRepo;
+import ru.itmo.highload.storoom.repositories.UnitRepo;
 import ru.itmo.highload.storoom.utils.Mapper;
 
 import java.util.List;
@@ -22,10 +25,11 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class LocationService {
     private final LocationRepo locationRepo;
+    private final UnitRepo unitRepo;
     private final OwnerRepo ownerRepo;
 
     public LocationEntity getEntityById(UUID id) {
-        return locationRepo.findById(id).orElseThrow(() -> new ResourceNotFoundException("lcoation " + id + " not found"));
+        return locationRepo.findById(id).orElseThrow(() -> new ResourceNotFoundException("Location", id));
     }
 
     public Page<DTOs.LocationReadDTO> getAll(Pageable pageable) {
@@ -37,11 +41,7 @@ public class LocationService {
         if (locationRepo.existsByAddress(dto.getAddress())) {
             throw new ResourceAlreadyExistsException();
         }
-        List<OwnerEntity> owners = ownerRepo.findByIdIn(dto.getOwnerIds())
-                .stream()
-                .map(OwnerEntity -> Optional.of(OwnerEntity))
-                .map(OwnerEntity -> OwnerEntity.orElseThrow(() -> new ResourceNotFoundException("owner " + OwnerEntity.get().getId() + " not found")))
-                .collect(Collectors.toList());
+        List<OwnerEntity> owners = getOwnersEntities(dto.getOwnerIds());
         return Mapper.toLocationDTO(locationRepo.save(Mapper.toLocationEntity(dto, owners)));
     }
 
@@ -49,20 +49,28 @@ public class LocationService {
         if (dto.getAddress() == null || dto.getAddress().isEmpty()) {
             throw new IllegalStateException("no name provided");
         }
-        LocationEntity locationEntity = locationRepo.findById(id).orElseThrow(() -> new ResourceNotFoundException("location " + id + " not found"));
+        LocationEntity locationEntity = getEntityById(id);
         locationEntity.setAddress(dto.getAddress());
         locationEntity.setLocationType(dto.getLocationType());
-        List<OwnerEntity> owners = ownerRepo.findByIdIn(dto.getOwnerIds())
-                .stream()
-                .map(OwnerEntity -> Optional.of(OwnerEntity))
-                .map(OwnerEntity -> OwnerEntity.orElseThrow(() -> new ResourceNotFoundException("owner " + id + " not found")))
-                .collect(Collectors.toList());
+        List<OwnerEntity> owners = getOwnersEntities(dto.getOwnerIds());
         locationEntity.setOwners(owners);
         return Mapper.toLocationDTO(locationRepo.save(locationEntity));
     }
 
     public void delete(UUID id) {
-        LocationEntity locationEntity = locationRepo.findById(id).orElseThrow(() -> new ResourceNotFoundException("location " + id + " not found"));
+        LocationEntity locationEntity = getEntityById(id);
+        List<UnitEntity> units = unitRepo.getAllByLocation(locationEntity);
+        if (units.stream().anyMatch(u -> u.getStatus().equals(UnitStatus.occupied))) {
+            throw new IllegalArgumentException("Close all units orders before deleting the location");
+        }
         locationRepo.delete(locationEntity);
+    }
+
+    private List<OwnerEntity> getOwnersEntities(List<UUID> ids) {
+        return ownerRepo.findByIdIn(ids)
+                .stream()
+                .map(Optional::of)
+                .map(ownerEntity -> ownerEntity.orElseThrow(() -> new ResourceNotFoundException("owner " + ownerEntity.get().getId() + " not found")))
+                .collect(Collectors.toList());
     }
 }
